backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "alembic>=1.17.2",
    "asyncpg>=0.30.0",
    "fastapi-users[oauth,sqlalchemy]>=15.0.1",
    "fastapi[standard]>=0.121.3",
    "httpx-oauth>=0.16.1",
    "loguru>=0.7.3",
    "pandas>=2.3.3",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "tqdm>=4.67.1",
    "openai-whisper>=20231117",
    "torch>=2.0.0",
    "spacy>=3.7.0",
    "openai>=1.12.0",
    "rapidfuzz>=3.0.0",
    "pl-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/pl_core_news_lg-3.8.0/pl_core_news_lg-3.8.0-py3-none-any.whl",
    "sentence-transformers>=5.2.0",
    "pydantic>=2.12.4",
    "llama-cpp-python>=0.3.16",
    "huggingface-hub>=0.36.0",
    "rank-bm25>=0.2.2",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }
torchaudio = { index = "pytorch" }
